copy compelete
