+++ client/go.mod
@@ -1,3 +1,5 @@
 module ciliumsr/client
 
 go 1.18
+
+require github.com/gorilla/websocket v1.5.0 // indirect
