+++ client/websocket/client.go
@@ -1,0 +1,1 @@
+package websocket
