+++ client/websocket/type.go
@@ -1,0 +1,6 @@
+package websocket
+
+type Message struct {
+	ip   int
+	tags string
+}
