+++ server/go.mod
@@ -1,3 +1,5 @@
 module ciliumsr/server
 
 go 1.18
+
+require github.com/gorilla/websocket v1.5.0
